data-journey/pyproject.toml
[project]
name = "data-journey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dataproc-spark-connect>=0.9.0",
    "db-dtypes>=1.4.4",
    "faker>=38.2.0",
    "fastavro>=1.12.1",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=3.6.0",
    "marimo>=0.18.1",
    "python-lsp-server>=1.13.2",
    "setuptools>=80.9.0",
    "sodapy>=2.2.0",
]
